% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPepSeq.R
\name{addPepSeq}
\alias{addPepSeq}
\title{Add/update column in blast table listing peptide names.}
\usage{
addPepSeq(blast, fasta)
}
\arguments{
\item{blast}{BLAST table to update with sequences.}

\item{fasta}{Fasta file with relevant peptide names and amino acid sequences.}
}
\description{
Add/update column in blast table listing peptide names.
}
