% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseIndex.R
\name{chooseIndex}
\alias{chooseIndex}
\title{Identifies which of a set of remaining peptides from a blast
alignment table should serve as the next index peptide.}
\usage{
chooseIndex(input, method = "min")
}
\arguments{
\item{input}{Data frame with BLAST alignments.}

\item{method}{Either "min" or "max": instructs whether chooseIndex should
suggest the peptide with the fewest or most alignments, respectively.}
}
\value{
Peptide name.
}
\description{
Identifies which of a set of remaining peptides from a blast
alignment table should serve as the next index peptide.
}
