% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleBLAST.R
\name{cycleBLAST}
\alias{cycleBLAST}
\title{cycleBLAST}
\usage{
cycleBLAST(data, pb, n, verbose = FALSE)
}
\arguments{
\item{data}{List of BLAST alignment table and fastra file to process.}

\item{pb}{txtProgressBar object to update}

\item{n}{Value to update txtProgressBar object.}

\item{verbose}{Logical whether or not to print status updates to console.}
}
\value{
List data containing modified BLAST alignment table and fasta file.
}
\description{
Repeatedly call epitopeBLAST to iterate through all index peptides
until all peptides are converted to trimmed epitopes.
}
