% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCleanup.R
\name{dbCleanup}
\alias{dbCleanup}
\title{dbCleanup
Directory management for epitopefindr}
\usage{
dbCleanup(relocate.input = FALSE, relocate.dir)
}
\arguments{
\item{relocate.input}{Logical whether to move input fasta file to new location.}

\item{relocate.dir}{Directory into which to move files if relocate.input == TRUE.}
}
\description{
dbCleanup
Directory management for epitopefindr
}
