% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decipherGaps.R
\name{decipherGaps}
\alias{decipherGaps}
\title{decipherGaps
Split gapped BLAST alignments into smaller ungapped alignments.}
\usage{
decipherGaps(blast)
}
\arguments{
\item{blast}{Input gapped BLAST alignment table to ungap.}
}
\description{
decipherGaps
Split gapped BLAST alignments into smaller ungapped alignments.
}
