% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epFind.R
\name{epFind}
\alias{epFind}
\title{epFind
All-in-one function to call the major steps of epitopefindr.}
\usage{
epFind(proj.id, e.thresh = 0.01, g.method = "any", aln.size = 7,
  autorun = TRUE, relocate.input = FALSE, relocate.dir)
}
\arguments{
\item{proj.id}{Name of peptide cohort. Output written to output/proj.id/.}

\item{e.thresh}{Maximum e value to allow from blast.alignments. Values
above 1 are not fully supported.}

\item{g.method}{Grouping method, either "any" or "all". See groupMSA
documentation for further details.}

\item{aln.size}{Shortest alignment to consider valid. Smaller alignments
are ignored and do not influence final epitopes.}

\item{autorun}{Logical whether to excute functions or just initialize BLAST.}

\item{relocate.input}{Logical whether to move input fasta file to new location.}

\item{relocate.dir}{Directory into which to move files if relocate.input == TRUE.}
}
\description{
epFind
All-in-one function to call the major steps of epitopefindr.
}
