% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epFind2.R
\name{epFind2}
\alias{epFind2}
\title{epFind2
New all-in-one function to call the major steps of epitopefindr.}
\usage{
epFind2(data = NULL, output.dir = NULL, e.thresh = 0.01,
  g.method = "any", aln.size = 7)
}
\arguments{
\item{data}{Biostrings::AAStringset input sequences to search for epitopes, or path to corresponding .fasta file.}

\item{output.dir}{Directory to which output files should be written.}

\item{e.thresh}{Maximum e-value to consider from BLASTp alignments of 'data'.}

\item{g.method}{Grouping method of alignments. Either 'any' or 'all'.}

\item{aln.size}{Minimum length of alignment to consider from BLASTp alignments of 'data'.}
}
\description{
epFind2
New all-in-one function to call the major steps of epitopefindr.
}
