% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitopeBLAST.R
\name{epitopeBLAST}
\alias{epitopeBLAST}
\title{epitopeBLAST
Identifies next index peptide, calls indexEpitopes to parse alignments to
that index peptide, and updates BLAST table accordingly.}
\usage{
epitopeBLAST(path, verbose = FALSE)
}
\arguments{
\item{path}{Path to current fasta file of sequences to process.}

\item{verbose}{Logical of whether or not to print status updates to console.}
}
\description{
epitopeBLAST
Identifies next index peptide, calls indexEpitopes to parse alignments to
that index peptide, and updates BLAST table accordingly.
}
