% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitopefindr.R
\docType{package}
\name{epitopefindr}
\alias{epitopefindr}
\alias{epitopefindr-package}
\title{epitopefindr: a package for identifying minimal overlaps from BLAST alignments.}
\description{
epitopefindr: a package for identifying minimal overlaps from BLAST alignments.
}
\section{Foo functions}{

The foo functions ...
}

