% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{epitopefindrEnv}
\alias{epitopefindrEnv}
\title{.onload
Set stringsAsFactors FALSE for epitopefindr}
\format{An object of class \code{environment} of length 0.}
\usage{
epitopefindrEnv
}
\description{
.onload
Set stringsAsFactors FALSE for epitopefindr
}
\keyword{datasets}
