% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epPB.R
\name{epPB}
\alias{epPB}
\title{epPB}
\usage{
epPB(min, max)
}
\arguments{
\item{min}{txtProgressBar minimum value}

\item{max}{txtProgressBar maximum value}
}
\description{
Short for epitope progress bar, epPB calls utils::txtProgressBar with some
standardized parameters to reduce redundancy.
}
