% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epSetupBLAST.R
\name{epSetupBLAST}
\alias{epSetupBLAST}
\title{epSetupBLAST
Run BLASTp on input peptides against each other & pre-process for use}
\usage{
epSetupBLAST()
}
\description{
epSetupBLAST
Run BLASTp on input peptides against each other & pre-process for use
}
