% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBLAST.R
\name{filterBLAST}
\alias{filterBLAST}
\title{filterBLAST}
\usage{
filterBLAST(blast)
}
\arguments{
\item{blast}{BLAST alignment table from which to remove singletons.}
}
\description{
Remove singletons, sequneces with no non-self alignments,
from BLAST alignment table.
}
