% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freadSF.R
\name{freadSF}
\alias{freadSF}
\title{freadSF
Custom data.table::fread wrapper with stringsasfactors set to FALSE}
\usage{
freadSF(input)
}
\arguments{
\item{input}{File to read into R.}
}
\description{
freadSF
Custom data.table::fread wrapper with stringsasfactors set to FALSE
}
