% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glGet.R
\name{glGet}
\alias{glGet}
\title{glGet
Retrieve data from global environment, e.g. from previous glAssign call.}
\usage{
glGet(name)
}
\arguments{
\item{name}{Varable to retrieve.}
}
\description{
glGet
Retrieve data from global environment, e.g. from previous glAssign call.
}
