% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glParamAssign.R
\name{glParamAssign}
\alias{glParamAssign}
\title{glParamAssign
Special case of glAssign to update a parameter in a table called gl.params.}
\usage{
glParamAssign(name, data)
}
\arguments{
\item{name}{ID of variable to populate first column "param" of gl.params}

\item{data}{ID of value to populate second column "value" of gl.params}
}
\description{
glParamAssign
Special case of glAssign to update a parameter in a table called gl.params.
}
