% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glParamGet.R
\name{glParamGet}
\alias{glParamGet}
\title{glParamGet}
\usage{
glParamGet(name)
}
\arguments{
\item{name}{Variable to retrieve.}
}
\description{
Special case of glGet to retrieve a value from object gl.params.
}
