% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMSA.R
\name{groupMSA}
\alias{groupMSA}
\title{groupMSA
Generate pdf multiple sequence alignment logos for identified aligning groups.}
\usage{
groupMSA(trim.groups = FALSE, make.png = FALSE)
}
\arguments{
\item{trim.groups}{Logical whether or not to apply msaTrim to edges of logos.}

\item{make.png}{Locial whether or not to convert PDF output to PNG.}
}
\description{
groupMSA
Generate pdf multiple sequence alignment logos for identified aligning groups.
}
