% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexEpitopes.R
\name{indexEpitopes}
\alias{indexEpitopes}
\title{indexEpitopes
For a specified index peptide, identify the maximal intervals that represent
the consensus overlap of reported BLAST alignments to that peptide. Update
the BLAST table accordingly.}
\usage{
indexEpitopes(blast.index)
}
\arguments{
\item{blast.index}{Name of index peptide to process.}
}
\description{
indexEpitopes
For a specified index peptide, identify the maximal intervals that represent
the consensus overlap of reported BLAST alignments to that peptide. Update
the BLAST table accordingly.
}
