% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOverlapping.R
\name{isOverlapping}
\alias{isOverlapping}
\title{isOverlapping
Checks whether the stard/end pos in pattern overlaps with the
start/end pos of each row in table by at least almin and
returns positions in table that overlap.}
\usage{
isOverlapping(pattern, table, almin = 7)
}
\arguments{
\item{pattern}{IR query pattern sequence positions.}

\item{table}{IR reference table positions to check for overlap with pattern.}

\item{almin}{Minimum allowable overlap considered valid.}
}
\value{
Positions in \code{table} that overlap.
}
\description{
isOverlapping
Checks whether the stard/end pos in pattern overlaps with the
start/end pos of each row in table by at least almin and
returns positions in table that overlap.
}
