% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lclhits}
\alias{lclhits}
\title{Virscan hits from patient LCL.}
\format{A Biostrings::AAStringSet object with 159 named peptide sequences.}
\source{
Larman Lab PhIP-Seq NGS.
}
\usage{
lclhits
}
\description{
Peptide seuqnces enriched in an interesting sample from PhIP-Seq.
}
\keyword{datasets}
