% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIR.R
\name{makeIR}
\alias{makeIR}
\title{makeIR}
\usage{
makeIR(df)
}
\arguments{
\item{df}{Data frame to convert to IR object.}
}
\description{
converts a two-column data frame to an IRanges object for overlap calculations.
}
