% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numAlignments.R
\name{numAlignments}
\alias{numAlignments}
\title{numAlignments
Add/update column to blast table listing # of alignments per query sequence.}
\usage{
numAlignments(input)
}
\arguments{
\item{input}{BLAST table to update with number of alignments per peptide.}
}
\description{
numAlignments
Add/update column to blast table listing # of alignments per query sequence.
}
