% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFiles.R
\name{outputFiles}
\alias{outputFiles}
\title{outputFiles
Generates some output files to report from the epitope search.}
\usage{
outputFiles()
}
\description{
outputFiles
Generates some output files to report from the epitope search.
}
