% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairwise_viral_hits}
\alias{pairwise_viral_hits}
\title{Virscan hits from AVARDA patient 1}
\format{A Biostrings::AAStringSet object with 159 named peptide sequences.}
\source{
Larman Lab PhIP-Seq NGS.
}
\usage{
pairwise_viral_hits
}
\description{
Peptide sequences enriched in an interesting sample from PhIP-Seq.
PhIP-Seq, or Phage Immunoprecipitation Sequencing, has been described by
Larman et. al.
}
\details{
https://www.biorxiv.org/content/early/2018/05/30/333625
}
\keyword{datasets}
