% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSignpost.R
\name{printSignpost}
\alias{printSignpost}
\title{printSignpost}
\usage{
printSignpost(step.num, ... = NA)
}
\arguments{
\item{step.num}{Numerif reference of which step to display status for.}

\item{...}{Extra parameters to past into step.num 0.}
}
\description{
Some status updates that can be printed if verbose == TRUE.
}
