% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBLASTaa.R
\name{queryBLASTaa}
\alias{queryBLASTaa}
\title{queryBLASTaa}
\usage{
queryBLASTaa(query, db)
}
\arguments{
\item{query}{Path to fasta file of query sequence.}

\item{db}{Path to fasta file of database sequences.}
}
\description{
BLAST query sequence against a database of sequences.
}
