% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSmallAln.R
\name{removeSmallAln}
\alias{removeSmallAln}
\title{removeSmallAln}
\usage{
removeSmallAln(blast, minsize = 7)
}
\arguments{
\item{blast}{Input BLAST table.}

\item{minsize}{Minimum length of alignment to keep.}
}
\description{
Delete rows in blast table for which the alignment is fewer than 7aa or
other custom threshold.
}
