% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameBLAST.R
\name{renameBLAST}
\alias{renameBLAST}
\title{Revert qID and sID columns of blast table back to original names based on naming map.}
\usage{
renameBLAST(blast, name.map)
}
\arguments{
\item{blast}{Blast table of alignments.}

\item{name.map}{Map of original and modified peptide names to convert.}
}
\description{
Revert qID and sID columns of blast table back to original names based on naming map.
}
