% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameFasta.R
\name{renameFasta}
\alias{renameFasta}
\title{Revert peptide names of a fasta file back to original names based on naming map.}
\usage{
renameFasta(fasta, name.map)
}
\arguments{
\item{fasta}{Blast table of alignments.}

\item{name.map}{Map of original and modified peptide names to convert.}
}
\description{
Revert peptide names of a fasta file back to original names based on naming map.
}
