% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfBLASTaa.R
\name{selfBLASTaa}
\alias{selfBLASTaa}
\title{Run BLASTp on amino acid sequences against themselves.}
\usage{
selfBLASTaa(path)
}
\arguments{
\item{path}{Path to fasta file of amino acid sequences to BLAST.}
}
\description{
Run BLASTp on amino acid sequences against themselves.
}
