% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshBLAST.R
\name{threshBLAST}
\alias{threshBLAST}
\title{Subset initial BLAST alignbment table to only include alignments below a
specified e-value threshold.}
\usage{
threshBLAST(blast, e.thresh)
}
\arguments{
\item{blast}{Initial BLAST alignment table.}

\item{e.thresh}{Maximum e-value threshold to retain.}
}
\description{
Subset initial BLAST alignbment table to only include alignments below a
specified e-value threshold.
}
