% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBLAST.R
\name{tidyBLAST}
\alias{tidyBLAST}
\title{Call a series of functions to tidy raw BLAST alignment table for use.}
\usage{
tidyBLAST(blast, fasta, aln.size, use.doParallel)
}
\arguments{
\item{blast}{Initial BLAST table to tidy.}

\item{fasta}{Path to fasta sequences to reference to annotate BLAST table}

\item{aln.size}{Minimum length of alignment to consider from BLASTp alignments of 'data'.
with peptide sequences.}

\item{use.doParallel}{Logical whether or not to use doParallel parallelization.}
}
\description{
Call a series of functions to tidy raw BLAST alignment table for use.
}
