% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimEpitopes.R
\name{trimEpitopes}
\alias{trimEpitopes}
\title{trimEpitopes
Update BLAST table positions to reflect smaller intervals when a subinterval
has been determined as the minimal overlap of alignments to a peptide.}
\usage{
trimEpitopes(path, tofilter = FALSE)
}
\arguments{
\item{path}{Path to current fasta file.}

\item{tofilter}{Binary whether or not to filter BLAST table entries.}
}
\description{
trimEpitopes
Update BLAST table positions to reflect smaller intervals when a subinterval
has been determined as the minimal overlap of alignments to a peptide.
}
