% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmergeFastaDuplicates.R
\name{unmergeFastaDuplicates}
\alias{unmergeFastaDuplicates}
\title{If a fasta entry has a concatenated name, e.g. from a previous call to
mergeFastaDuplicates, then separate the concatenanted name and make
a separate entry for each.}
\usage{
unmergeFastaDuplicates(input, sep = "__________")
}
\arguments{
\item{input}{XStringSet object with concatenated names.}

\item{sep}{Delimiter to search for and parse.}
}
\value{
XStringSet object modified so that a single entries with delimited
names are replaced with multiple entries with identical sequences and
unmerged names.
}
\description{
If a fasta entry has a concatenated name, e.g. from a previous call to
mergeFastaDuplicates, then separate the concatenanted name and make
a separate entry for each.
}
