% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFastaAA.R
\name{writeFastaAA}
\alias{writeFastaAA}
\title{Takes data frame with $Seq and $ID or AAStringSet and writes fasta file}
\usage{
writeFastaAA(input, fpath)
}
\arguments{
\item{input}{Data frame or AAStringSet to write to fasta file.}

\item{fpath}{Directory path to write fasta file to.}
}
\description{
Takes data frame with $Seq and $ID or AAStringSet and writes fasta file
}
