% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleBLAST.R
\name{cycleBLAST}
\alias{cycleBLAST}
\title{cycleBLAST
Repeatedly call epitopeBLAST to iterate through all index peptides
until all peptides are converted to trimmed epitopes.}
\usage{
cycleBLAST(path, pb, n, verbose = FALSE)
}
\arguments{
\item{path}{Path to current fasta file.}

\item{pb}{txtProgressBar object to update}

\item{n}{Value to update txtProgressBar object.}

\item{verbose}{Logical whether or not to print status updates to console.}
}
\description{
cycleBLAST
Repeatedly call epitopeBLAST to iterate through all index peptides
until all peptides are converted to trimmed epitopes.
}
