% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decipherGaps.R
\name{decipherGaps}
\alias{decipherGaps}
\title{Split gapped BLAST alignments into smaller ungapped alignments.}
\usage{
decipherGaps(blast, use.doParallel)
}
\arguments{
\item{blast}{Input gapped BLAST alignment table to ungap.}

\item{use.doParallel}{Logical whether or not to use doParallel parallelization.}
}
\description{
Split gapped BLAST alignments into smaller ungapped alignments.
}
