% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epfind.R
\name{epfind}
\alias{epfind}
\alias{epFind2}
\title{All-in-one function to call the major steps of epitopefindr.}
\usage{
epfind(data = NULL, output.dir = NULL, e.thresh = 0.01,
  g.method = "any", aln.size = 7, min.groupsize = 2,
  min.consensus.pos = 1, consensus.thresh = c(75, 50),
  peptide.nchar = 50, msa.width = "dynamic", verbose = TRUE,
  pdflatex = TRUE, pdftk = TRUE, pdfuniter = TRUE,
  make.png = FALSE, name.msa = "msa.pdf",
  name.alignments = "finalAlignments.csv",
  name.epitopekey = "epitopeKey.csv",
  name.epitopesum = "epitopeSummary.csv", use.doParallel = FALSE)

epFind2(data = NULL, output.dir = NULL, e.thresh = 0.01,
  g.method = "any", aln.size = 7, min.groupsize = 2,
  min.consensus.pos = 1, consensus.thresh = c(75, 50),
  peptide.nchar = 50, msa.width = "dynamic", verbose = TRUE,
  pdflatex = TRUE, pdftk = TRUE, pdfuniter = TRUE,
  make.png = FALSE, name.msa = "msa.pdf",
  name.alignments = "finalAlignments.csv",
  name.epitopekey = "epitopeKey.csv",
  name.epitopesum = "epitopeSummary.csv", use.doParallel = FALSE)
}
\arguments{
\item{data}{Biostrings::AAStringset input sequences to search for epitopes, or path to corresponding .fasta file.}

\item{output.dir}{Directory to which output files should be written.}

\item{e.thresh}{Maximum e-value to consider from BLASTp alignments of 'data'.}

\item{g.method}{Grouping method of alignments. Either 'any' or 'all'. See ?indexGroups}

\item{aln.size}{Minimum length of alignment to consider from BLASTp alignments of 'data'.}

\item{min.groupsize}{Minimum number of peptides per group to require in order to print a group.}

\item{min.consensus.pos}{Minimum number of amino acid consensus positions required in order to print a group.}

\item{consensus.thresh}{Two decreasing numeric values of upper and lower thresholds for sequence consensus.}

\item{peptide.nchar}{Maximum of character from peptide name to use in msa output. Default 50. Starts from left.}

\item{msa.width}{Controls whether or not MSA images have fixed or dynamic width. By default, msa.width is set to "dynamic", which causes the document dimentions of the resultant image to be calculated based on the lentht of the peptide name and the number of amino acids in the sequence alignment. If msa.width is instead set to a numeric, then an MSA will be printed with a fixed with that number of inches. With 50-character peptide.nchar and a maximum expected sequence alignment of 45 positions, an msa.width of 12 is more than sufficient.}

\item{verbose}{Logical to print progress updates.}

\item{pdflatex}{Logical whether or not to produce PDF LaTeX figures using pdflatex}

\item{pdftk}{Logical whether or not to merge msa pdfs using staplr and pdftk}

\item{pdfuniter}{Logical whether or not to merge msa pdfs using staplr and pdfuniter}

\item{make.png}{Locial whether or not to convert PDF output to PNG.}

\item{name.msa}{Filename for output merged pdf of msa logos.}

\item{name.alignments}{Filename for output spreadsheet of peptide alignments.}

\item{name.epitopekey}{Filename for output spreadhseet of epitopes per peptide.}

\item{name.epitopesum}{Filename for output summary sheet of epitopes.}

\item{use.doParallel}{Logical whether or not to use doParallel parallelization.}
}
\description{
All-in-one function to call the major steps of epitopefindr.
}
