% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epFind2.R
\name{epFind2}
\alias{epFind2}
\title{epFind2}
\usage{
epFind2(data = NULL, output.dir = NULL, e.thresh = 0.01,
  g.method = "any", aln.size = 7, min.groupsize = 2,
  min.consensus.pos = 1, consensus.thresh = c(75, 50),
  verbose = TRUE, pdflatex = TRUE, pdftk = TRUE,
  name.msa = "msa.pdf", name.alignments = "finalAlignments.csv",
  name.epitopekey = "epitopeKey.csv",
  name.epitopesum = "epitopeSummary.csv")
}
\arguments{
\item{data}{Biostrings::AAStringset input sequences to search for epitopes, or path to corresponding .fasta file.}

\item{output.dir}{Directory to which output files should be written.}

\item{e.thresh}{Maximum e-value to consider from BLASTp alignments of 'data'.}

\item{g.method}{Grouping method of alignments. Either 'any' or 'all'. See ?indexGroups}

\item{aln.size}{Minimum length of alignment to consider from BLASTp alignments of 'data'.}

\item{min.groupsize}{Minimum number of peptides per group to require in order to print a group.}

\item{min.consensus.pos}{Minimum number of amino acid consensus positions required in order to print a group.}

\item{consensus.thresh}{Two decreasing numeric values of upper and lower thresholds for sequence consensus.}

\item{verbose}{Logical to print progress updates.}

\item{pdflatex}{Logical whether or not to produce PDF LaTeX figures using pdflatex}

\item{pdftk}{Logical whether or not to merge msa pdfs using staplr and pdftk}

\item{name.msa}{Filename for output merged pdf of msa logos.}

\item{name.alignments}{Filename for output spreadsheet of peptide alignments.}

\item{name.epitopekey}{Filename for output spreadhseet of epitopes per peptide.}

\item{name.epitopesum}{Filename for output summary sheet of epitopes.}
}
\description{
New all-in-one function to call the major steps of epitopefindr.
}
