% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitopeBLAST.R
\name{epitopeBLAST}
\alias{epitopeBLAST}
\title{Identifies next index peptide, calls indexEpitopes to parse alignments to
that index peptide, and updates BLAST table accordingly.}
\usage{
epitopeBLAST(data, aln.size)
}
\arguments{
\item{data}{List with BLAST alignment table and fasta file of peptides.}

\item{aln.size}{Minimum length of alignment to consider from BLASTp alignments of 'data'.}
}
\description{
Identifies next index peptide, calls indexEpitopes to parse alignments to
that index peptide, and updates BLAST table accordingly.
}
