% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitopeBLAST.R
\name{epitopeBLAST}
\alias{epitopeBLAST}
\title{epitopeBLAST
Identifies next index peptide, calls indexEpitopes to parse alignments to
that index peptide, and updates BLAST table accordingly.}
\usage{
epitopeBLAST(data)
}
\arguments{
\item{data}{List with BLAST alignment table and fasta file of peptides.}

\item{fasta}{Fasta file of sequences to process.}
}
\description{
epitopeBLAST
Identifies next index peptide, calls indexEpitopes to parse alignments to
that index peptide, and updates BLAST table accordingly.
}
