% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epSetupDirectory.R
\name{epSetupDirectory}
\alias{epSetupDirectory}
\title{epSetupDirectory
Setup epitopefinder directory paths}
\usage{
epSetupDirectory(proj.id, e.thresh, g.method)
}
\description{
epSetupDirectory
Setup epitopefinder directory paths
}
