% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glAssign.R
\name{glAssign}
\alias{glAssign}
\title{glAssign}
\usage{
glAssign(name, data)
}
\arguments{
\item{name}{Identifier for object.}

\item{data}{Corresponding information.}
}
\description{
Quick wrapper function for assigning data to global environment
}
