% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMSA.R
\name{groupMSA}
\alias{groupMSA}
\title{Generate pdf multiple sequence alignment logos for identified aligning groups.}
\usage{
groupMSA(groups, mpath = "intermediate_files/msa/", min.groupsize = 2,
  min.consensus.pos = 1, consensus.thresh = c(75, 50),
  pdflatex = TRUE, pdftk = TRUE, trim.groups = FALSE,
  make.png = FALSE)
}
\arguments{
\item{groups}{Table of final peptides and pre-calculated groupings.}

\item{mpath}{Directory to write sequence alignment files.}

\item{min.groupsize}{Minimum number of peptides per group to require in order to print a group.}

\item{min.consensus.pos}{Minimum number of amino acid consensus positions required in order to print a group.}

\item{consensus.thresh}{Two decreasing numeric values of upper and lower thresholds for sequence consensus.}

\item{pdflatex}{Logical whether or not to produce PDF LaTeX figures using pdflatex}

\item{pdftk}{Logical whether or not to use staplr and pdftk to merge individual msa pdfs.}

\item{trim.groups}{Logical whether or not to apply msaTrim to edges of logos. Not implemented.}

\item{make.png}{Locial whether or not to convert PDF output to PNG.}
}
\description{
Generate pdf multiple sequence alignment logos for identified aligning groups.
}
