% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexGroups.R
\name{indexGroups}
\alias{indexGroups}
\title{indexGroups
Group trimmed sequences into aligning groups}
\usage{
indexGroups(blast, fasta, mode = "any")
}
\arguments{
\item{blast}{BLAST alignment table to process.}

\item{fasta}{Epitope sequence AAStringset object to process.}

\item{mode}{Grouping method. "any" creates a smaller number of groups, with
individual groups tending to have more members, such that in each group,
each member must align with at least 1 "any" of the other members. "all"
creates a larger number of groups, with individual groups tending to have
fewer members, wuch taht in each group, each member must align with "all"
other members.}
}
\description{
indexGroups
Group trimmed sequences into aligning groups
}
