% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFastaDuplicates.R
\name{mergeFastaDuplicates}
\alias{mergeFastaDuplicates}
\title{If two fasta entries have the same sequence, keep only one copy of the
sequence but concatenante the name to be Name1__Name2; __ can be replaced
with delimiter of choice.}
\usage{
mergeFastaDuplicates(input, sep = "__________")
}
\arguments{
\item{input}{XStringSet object to collapse.}

\item{sep}{Delimiter to place between names.}
}
\description{
If two fasta entries have the same sequence, keep only one copy of the
sequence but concatenante the name to be Name1__Name2; __ can be replaced
with delimiter of choice.
}
