% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeBLAST.R
\name{organizeBLAST}
\alias{organizeBLAST}
\title{organizeBLAST}
\usage{
organizeBLAST(input)
}
\arguments{
\item{input}{Input default BLAST table.}
}
\description{
Rename unwieldy columns from rBLAST output table.
}
