% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputTable.R
\name{outputTable}
\alias{outputTable}
\title{outputTable
Generates spreadsheet of reportable data on identified minimal alignments.}
\usage{
outputTable()
}
\description{
outputTable
Generates spreadsheet of reportable data on identified minimal alignments.
}
