% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbCycleBLAST.R
\name{pbCycleBLAST}
\alias{pbCycleBLAST}
\title{pbCycleBLAST
Initialize a txtProgressBar object with which to call cycleBLAST.}
\usage{
pbCycleBLAST(path, ncycles = "max")
}
\arguments{
\item{path}{Path to current fasta file to pass to cycleBLAST.}

\item{ncycles}{Number of cycles of cycleBLAST to perform.}
}
\description{
pbCycleBLAST
Initialize a txtProgressBar object with which to call cycleBLAST.
}
