% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBLAST.R
\name{prepareBLAST}
\alias{prepareBLAST}
\title{prepareBLAST
Call series of functions to standardize BLAST table for computation.}
\usage{
prepareBLAST(blast, fasta, tofilter = TRUE)
}
\arguments{
\item{blast}{Input tidied BLAST table.}

\item{fasta}{AAstringset object of fasta sequences.}

\item{tofilter}{Logical whether or not to call filterBLAST, which is only
necessary during initial setup.}
}
\description{
prepareBLAST
Call series of functions to standardize BLAST table for computation.
}
