% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBLAST.R
\name{tidyBLAST}
\alias{tidyBLAST}
\title{tidyBLAST}
\usage{
tidyBLAST(blast, fasta)
}
\arguments{
\item{blast}{Initial BLAST table to tidy.}

\item{fasta}{Path to fasta sequences to reference to annotate BLAST table
with peptide sequences.}
}
\description{
Call a series of functions to tidy raw BLAST alignment table for use.
}
