% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFasta.R
\name{tidyFasta}
\alias{tidyFasta}
\title{tidyFasta
remove filler sequences after "." stop codon and cterminal "*"}
\usage{
tidyFasta(input)
}
\arguments{
\item{input}{Untidy fasta.}
}
\description{
tidyFasta
remove filler sequences after "." stop codon and cterminal "*"
}
